test_requirements.txt
astroid==2.0.4
black==18.9b0
bpython
codecov
django-debug-toolbar
factory_boy
ipdb
nplusone>=0.8.1
pdbpp
pip-tools
pylint==2.1.1
pylint-django==2.0.2
pytest
pytest-cov>=2.6.1
pytest-django
pytest-pep8
pytest-mock
pytest-pylint==0.12.3
pytest-lazy-fixture==0.4.2
safety
semantic-version
